pyproject.toml
[tool.poetry]
name = "ragwork"
version = "0.1.0"
description = ""
authors = ["anualiassynbrains <anu.alias@synbrains.ai>"]
readme = "README.md"
packages = [{include = "ragwork", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
sentence-transformers = "^5.0.0"
strawberry-graphql = "^0.275.5"
langchain-groq = "^0.3.5"
python-dotenv = "^1.1.1"
scikit-learn = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
